Change interface to separate monetization and license
