+++ components/plan-designer/MonetizationMethod.vue
@@ -1,15 +1,12 @@
 <script setup lang="ts">
-import dayjs from 'dayjs';
 import { useI18n } from 'vue-i18n';
 import { z } from 'zod';
 
 import { licenses } from '~/constants/licenses';
 import type CardSelection from '~/interfaces/card-selection';
-import { DownloadFrequency } from '~/interfaces/download-frequency.enum';
 import { SubscriptionFrequency } from '~/interfaces/subscription-frequency.enum';
 
 //TODO: Get this list of countries from somewhere like an API call
-const listOfCountries = ['Greece', 'Cyprus', 'Italy', 'Germany', 'France'];
 
 const { showErrorMessage } = useAlertMessage();
 const { t } = useI18n();
@@ -154,31 +151,6 @@ const monetizationSelections: CardSelection[] = [
     },
 ];
 
-const licenseSelections = computed(() => (isFree.value ? Object.values(licenses) : [licenses.pistis]));
-
-const transferableSelections = [
-    {
-        label: t('data.designer.transferable'),
-        value: 'transferable',
-    },
-    {
-        label: t('data.designer.nonTransferable'),
-        value: 'non-transferable',
-    },
-    {
-        label: t('data.designer.subLicensable'),
-        value: 'sub-licensable',
-    },
-];
-
-const limitFrequencySelections = computed(() => [
-    { title: t('perHour'), value: DownloadFrequency.HOUR as string },
-    { title: t('perDay'), value: DownloadFrequency.DAY as string },
-    { title: t('perWeek'), value: DownloadFrequency.WEEK as string },
-    { title: t('perMonth'), value: DownloadFrequency.MONTH as string },
-    { title: t('perYear'), value: DownloadFrequency.YEAR as string },
-]);
-
 const emit = defineEmits([
     'update:monetization-details-prop',
     'update:is-free',
@@ -200,34 +172,6 @@ const updateFree = (value: boolean) => {
     } else {
         monetizationDetails.value.price = undefined;
     }
-};
-
-const updateWorldwide = (value: boolean) => {
-    isWorldwide.value = value;
-    emit('update:is-worldwide', value);
-    if (isWorldwide.value) {
-        monetizationDetails.value.region = '';
-    }
-};
-
-const updatePerpetual = (value: boolean) => {
-    isPerpetual.value = value;
-    emit('update:is-perpetual', value);
-    if (isPerpetual.value) {
-        monetizationDetails.value.termDate = undefined;
-    }
-};
-
-const updatePersonal = (value: boolean) => {
-    hasPersonalData.value = value;
-    emit('update:has-personal-data', value);
-    if (!hasPersonalData.value) {
-        monetizationDetails.value.personalDataTerms = '';
-    }
-};
-
-const updateContractTerms = (value: string) => {
-    monetizationDetails.value.contractTerms = value;
 };
 
 const handleMonetizationClick = (value: string) => {
@@ -357,279 +301,6 @@ async function onSubmit(): Promise<void> {
                                             />
                                         </UFormGroup>
                                     </div>
-                                    <UFormGroup
-                                        :label="$t('license')"
-                                        required
-                                        name="license"
-                                        class="flex-1 text-gray-200"
-                                    >
-                                        <USelectMenu
-                                            v-model="monetizationDetails.license"
-                                            :ui="{
-                                                option: { active: 'text-gray-200' },
-                                                input: 'placeholder:text-gray-200 text-gray-200',
-                                                button: 'text-gray-200',
-                                            }"
-                                            :placeholder="$t('data.designer.selectLicense')"
-                                            :options="licenseSelections"
-                                            ><template #label>
-                                                <span v-if="monetizationDetails.license" class="truncate">{{
-                                                    monetizationDetails.license
-                                                }}</span>
-                                                <span v-else class="text-gray-400">Select license</span>
-                                            </template></USelectMenu
-                                        >
-                                    </UFormGroup>
-                                </div>
-                                <div class="flex flex-row gap-4">
-                                    <UFormGroup
-                                        :label="$t('data.designer.downloadLimit')"
-                                        required
-                                        name="limitNumber"
-                                        class="flex-1"
-                                    >
-                                        <UInput
-                                            v-model.number="monetizationDetails.limitNumber"
-                                            :placeholder="$t('data.designer.downloadLimitPH')"
-                                            type="numeric"
-                                        >
-                                            <template #trailing>
-                                                <span class="text-gray-500 text-xs">{{ $t('times') }}</span>
-                                            </template>
-                                        </UInput>
-                                    </UFormGroup>
-                                    <UFormGroup :label="$t('frequency')" required name="limitFrequency" class="flex-1">
-                                        <USelectMenu
-                                            v-model="monetizationDetails.limitFrequency"
-                                            :placeholder="$t('data.designer.selectFrequency')"
-                                            :options="limitFrequencySelections"
-                                            value-attribute="value"
-                                            option-attribute="title"
-                                            ><template #label>
-                                                <span v-if="monetizationDetails.limitFrequency" class="truncate">{{
-                                                    monetizationDetails.limitFrequency
-                                                }}</span>
-                                                <span v-else class="text-gray-400">Select a frequency</span>
-                                            </template></USelectMenu
-                                        >
-                                    </UFormGroup>
-                                </div>
-                                <div v-if="monetizationDetails.license === licenses.pistis" class="flex flex-col gap-4">
-                                    <div class="flex flex-row gap-4">
-                                        <div class="flex flex-1 gap-4">
-                                            <UFormGroup :label="$t('exclusive')" name="isExclusive">
-                                                <UCheckbox
-                                                    v-model="monetizationDetails.isExclusive"
-                                                    name="isExclusive"
-                                                    class="mt-2.5 -ml-1 justify-center"
-                                                />
-                                            </UFormGroup>
-                                            <UFormGroup
-                                                :label="$t('data.designer.availability')"
-                                                name="region"
-                                                :required="!isWorldwide"
-                                                class="w-full"
-                                            >
-                                                <!-- <UInput
-                                                    v-model.number="monetizationDetails.region"
-                                                    :class="isWorldwide ? 'opacity-50' : ''"
-                                                    :disabled="isWorldwide"
-                                                    :placeholder="$t('data.designer.regionCountry')"
-                                                    type="numeric"
-                                                    class="w-full"
-                                                >
-                                                </UInput> -->
-                                                <USelectMenu
-                                                    v-model="monetizationDetails.region"
-                                                    :ui="{
-                                                        option: { active: 'text-gray-200' },
-                                                        input: 'placeholder:text-gray-200 text-gray-200',
-                                                        button: 'text-gray-200',
-                                                    }"
-                                                    :options="listOfCountries"
-                                                    :placeholder="$t('data.designer.regionCountry')"
-                                                ></USelectMenu>
-
-                                                <template #error="{ error }">
-                                                    <span
-                                                        :class="[
-                                                            error
-                                                                ? 'text-red-500 dark:text-red-400'
-                                                                : 'text-primary-500 dark:text-primary-400',
-                                                        ]"
-                                                    >
-                                                        {{ isWorldwide ? '' : error }}
-                                                    </span>
-                                                </template>
-                                            </UFormGroup>
-                                            <UFormGroup :label="$t('data.designer.worldwide')">
-                                                <UCheckbox
-                                                    :model-value="isWorldwide"
-                                                    class="mt-2.5 justify-center"
-                                                    @update:model-value="(value: boolean) => updateWorldwide(value)"
-                                                />
-                                            </UFormGroup>
-                                        </div>
-                                        <div class="flex flex-1 gap-4">
-                                            <UFormGroup
-                                                :label="$t('data.designer.transferable')"
-                                                required
-                                                name="transferable"
-                                                class="flex-1 text-gray-200"
-                                            >
-                                                <USelectMenu
-                                                    v-model="monetizationDetails.transferable"
-                                                    :ui="{
-                                                        option: { active: 'text-gray-200' },
-                                                        input: 'placeholder:text-gray-200 text-gray-200',
-                                                        button: 'text-gray-200',
-                                                    }"
-                                                    :placeholder="$t('data.designer.transferable')"
-                                                    :options="transferableSelections"
-                                                    value-attribute="value"
-                                                    option-attribute="label"
-                                                ></USelectMenu>
-                                            </UFormGroup>
-                                            <UFormGroup
-                                                :label="$t('data.designer.termDate')"
-                                                :required="!isPerpetual"
-                                                name="termDate"
-                                                class="text-gray-200"
-                                            >
-                                                <UPopover :popper="{ placement: 'bottom-start' }">
-                                                    <UButton
-                                                        color="white"
-                                                        icon="i-heroicons-calendar-days-20-solid"
-                                                        :label="
-                                                            monetizationDetails.termDate
-                                                                ? dayjs(monetizationDetails.termDate).format(
-                                                                      'DD MMMM YYYY',
-                                                                  )
-                                                                : isPerpetual
-                                                                  ? t('data.designer.licenseIsPerpetual')
-                                                                  : t('data.designer.pleaseSelectDate')
-                                                        "
-                                                        :disabled="isPerpetual"
-                                                        :class="[
-                                                            isPerpetual ? 'opacity-50' : '',
-                                                            monetizationDetails.termDate
-                                                                ? 'text-gray-700'
-                                                                : 'text-gray-400',
-                                                        ]"
-                                                    />
-                                                    <template #panel="{ close }">
-                                                        <DatePicker
-                                                            v-model="monetizationDetails.termDate"
-                                                            is-required
-                                                            @close="close"
-                                                        />
-                                                    </template>
-                                                </UPopover>
-                                                <template #error="{ error }">
-                                                    <span
-                                                        :class="[
-                                                            error
-                                                                ? 'text-red-500 dark:text-red-400'
-                                                                : 'text-primary-500 dark:text-primary-400',
-                                                        ]"
-                                                    >
-                                                        {{ isPerpetual || monetizationDetails.termDate ? '' : error }}
-                                                    </span>
-                                                </template>
-                                            </UFormGroup>
-                                            <UFormGroup :label="$t('data.designer.perpetual')">
-                                                <UCheckbox
-                                                    :model-value="isPerpetual"
-                                                    class="mt-2.5 justify-center"
-                                                    @update:model-value="(value: boolean) => updatePerpetual(value)"
-                                                />
-                                            </UFormGroup>
-                                        </div>
-                                    </div>
-                                    <UFormGroup :label="$t('termsConditions')" name="extraTerms">
-                                        <UTextarea
-                                            v-model="monetizationDetails.extraTerms"
-                                            :rows="4"
-                                            :placeholder="$t('data.designer.typeTerms')"
-                                            resize
-                                        />
-                                    </UFormGroup>
-                                    <UFormGroup
-                                        :label="$t('data.designer.additionalRenewalTerms')"
-                                        name="additionalRenewalTerms"
-                                    >
-                                        <UTextarea
-                                            v-model="monetizationDetails.additionalRenewalTerms"
-                                            :rows="4"
-                                            :placeholder="$t('data.designer.additionalRenewalTerms')"
-                                            resize
-                                        />
-                                    </UFormGroup>
-                                    <div class="flex flex-row gap-4">
-                                        <div class="flex-1 flex gap-4">
-                                            <UFormGroup
-                                                :label="$t('data.designer.noticeForNonRenewal')"
-                                                class="flex-1"
-                                                required
-                                                name="nonRenewalDays"
-                                            >
-                                                <UInput
-                                                    v-model.number="monetizationDetails.nonRenewalDays"
-                                                    :placeholder="$t('data.designer.noticeForNonRenewal')"
-                                                    type="numeric"
-                                                >
-                                                    <template #trailing>
-                                                        <span class="text-gray-500 text-xs">{{ t('days') }}</span>
-                                                    </template>
-                                                </UInput>
-                                            </UFormGroup>
-                                        </div>
-                                        <div class="flex-1 flex gap-4">
-                                            <UFormGroup
-                                                :label="$t('data.designer.maximumDaysContractBreach')"
-                                                class="flex-1"
-                                                required
-                                                name="contractBreachDays"
-                                            >
-                                                <UInput
-                                                    v-model.number="monetizationDetails.contractBreachDays"
-                                                    :placeholder="$t('data.designer.maximumDaysContractBreach')"
-                                                    type="numeric"
-                                                >
-                                                    <template #trailing>
-                                                        <span class="text-gray-500 text-xs">{{ t('days') }}</span>
-                                                    </template>
-                                                </UInput>
-                                            </UFormGroup>
-                                        </div>
-                                    </div>
-                                    <div class="relative">
-                                        <div
-                                            class="absolute right-0 top-0 text-sm flex gap-4 items-center font-semibold"
-                                        >
-                                            <span class="text-gray-700">
-                                                {{ t('data.designer.hasPersonalData') }}
-                                            </span>
-                                            <UCheckbox
-                                                :model-value="hasPersonalData"
-                                                class="justify-center"
-                                                @update:model-value="(value: boolean) => updatePersonal(value)"
-                                            />
-                                        </div>
-                                        <UFormGroup
-                                            :label="$t('data.designer.personalDataTerms')"
-                                            name="personalDataTerms"
-                                            :required="hasPersonalData"
-                                        >
-                                            <UTextarea
-                                                v-model="monetizationDetails.personalDataTerms"
-                                                :rows="4"
-                                                :placeholder="$t('data.designer.personalDataTerms')"
-                                                resize
-                                                :disabled="!hasPersonalData"
-                                            />
-                                        </UFormGroup>
-                                    </div>
                                 </div>
                             </div>
                         </template>
@@ -637,7 +308,7 @@ async function onSubmit(): Promise<void> {
                         <template v-if="monetizationDetails.type === 'subscription'">
                             <div class="flex flex-col space-y-5">
                                 <div class="flex flex-row gap-4">
-                                    <div class="flex gap-4 w-1/2">
+                                    <div class="flex gap-4 w-full">
                                         <UFormGroup
                                             :label="$t('data.designer.subscriptionFrequency')"
                                             required
@@ -708,259 +379,6 @@ async function onSubmit(): Promise<void> {
                                             </UFormGroup>
                                         </div>
                                     </div>
-                                    <UFormGroup :label="$t('license')" required class="w-1/2" name="license">
-                                        <USelectMenu
-                                            v-model="monetizationDetails.license"
-                                            :class="'gray'"
-                                            :placeholder="$t('data.designer.selectLicense')"
-                                            :options="licenseSelections"
-                                            ><template #label>
-                                                <span v-if="monetizationDetails.license" class="truncate">{{
-                                                    monetizationDetails.license
-                                                }}</span>
-                                                <span v-else class="text-gray-400">Select a license</span>
-                                            </template></USelectMenu
-                                        >
-                                    </UFormGroup>
-                                </div>
-
-                                <div class="flex flex-row gap-4">
-                                    <UFormGroup
-                                        :label="$t('data.designer.downloadLimit')"
-                                        class="flex-1"
-                                        required
-                                        name="limitNumber"
-                                    >
-                                        <UInput
-                                            v-model.number="monetizationDetails.limitNumber"
-                                            :placeholder="$t('data.designer.downloadLimitPH')"
-                                            type="numeric"
-                                        >
-                                            <template #trailing>
-                                                <span class="text-gray-500 text-xs">times</span>
-                                            </template>
-                                        </UInput>
-                                    </UFormGroup>
-                                    <UFormGroup :label="$t('frequency')" required name="limitFrequency" class="flex-1">
-                                        <USelectMenu
-                                            v-model="monetizationDetails.limitFrequency"
-                                            :placeholder="$t('data.designer.selectFrequency')"
-                                            :options="limitFrequencySelections"
-                                            value-attribute="value"
-                                            option-attribute="title"
-                                            ><template #label>
-                                                <span v-if="monetizationDetails.limitFrequency" class="truncate">{{
-                                                    monetizationDetails.limitFrequency
-                                                }}</span>
-                                                <span v-else class="text-gray-400">Select a frequency</span>
-                                            </template></USelectMenu
-                                        >
-                                    </UFormGroup>
-                                </div>
-                                <div v-if="monetizationDetails.license === licenses.pistis" class="flex flex-col gap-4">
-                                    <div class="flex flex-row gap-4">
-                                        <div class="flex flex-1 gap-4">
-                                            <UFormGroup :label="$t('exclusive')" name="isExclusive">
-                                                <UCheckbox
-                                                    v-model="monetizationDetails.isExclusive"
-                                                    name="isExclusive"
-                                                    class="mt-2.5 -ml-1 justify-center"
-                                                />
-                                            </UFormGroup>
-                                            <UFormGroup
-                                                :label="$t('data.designer.availability')"
-                                                name="region"
-                                                :required="!isWorldwide"
-                                                class="w-full"
-                                            >
-                                                <!-- <UInput
-                                                    v-model.number="monetizationDetails.region"
-                                                    :class="isWorldwide ? 'opacity-50' : ''"
-                                                    :disabled="isWorldwide"
-                                                    :placeholder="$t('data.designer.regionCountry')"
-                                                    type="numeric"
-                                                    class="w-full"
-                                                >
-                                                </UInput> -->
-                                                <USelectMenu
-                                                    v-model="monetizationDetails.region"
-                                                    :ui="{
-                                                        option: { active: 'text-gray-200' },
-                                                        input: 'placeholder:text-gray-200 text-gray-200',
-                                                        button: 'text-gray-200',
-                                                    }"
-                                                    :options="listOfCountries"
-                                                    :placeholder="$t('data.designer.regionCountry')"
-                                                ></USelectMenu>
-                                            </UFormGroup>
-                                            <UFormGroup :label="$t('data.designer.worldwide')">
-                                                <UCheckbox
-                                                    :model-value="isWorldwide"
-                                                    class="mt-2.5 justify-center"
-                                                    @update:model-value="(value: boolean) => updateWorldwide(value)"
-                                                />
-                                            </UFormGroup>
-                                        </div>
-                                        <div class="flex flex-1 gap-4">
-                                            <UFormGroup
-                                                :label="$t('data.designer.transferable')"
-                                                required
-                                                name="transferable"
-                                                class="flex-1 text-gray-200"
-                                            >
-                                                <USelectMenu
-                                                    v-model="monetizationDetails.transferable"
-                                                    :ui="{
-                                                        option: { active: 'text-gray-200' },
-                                                        input: 'placeholder:text-gray-200 text-gray-200',
-                                                        button: 'text-gray-200',
-                                                    }"
-                                                    :placeholder="$t('data.designer.transferable')"
-                                                    :options="transferableSelections"
-                                                    value-attribute="value"
-                                                    option-attribute="label"
-                                                ></USelectMenu>
-                                            </UFormGroup>
-                                            <UFormGroup
-                                                :label="$t('data.designer.termDate')"
-                                                :required="!isPerpetual"
-                                                name="termDate"
-                                                class="text-gray-200"
-                                            >
-                                                <UPopover :popper="{ placement: 'bottom-start' }">
-                                                    <UButton
-                                                        color="white"
-                                                        icon="i-heroicons-calendar-days-20-solid"
-                                                        :label="
-                                                            monetizationDetails.termDate
-                                                                ? dayjs(monetizationDetails.termDate).format(
-                                                                      'DD MMMM YYYY',
-                                                                  )
-                                                                : isPerpetual
-                                                                  ? t('data.designer.licenseIsPerpetual')
-                                                                  : t('data.designer.pleaseSelectDate')
-                                                        "
-                                                        :disabled="isPerpetual"
-                                                        :class="[
-                                                            isPerpetual ? 'opacity-50' : '',
-                                                            monetizationDetails.termDate
-                                                                ? 'text-gray-700'
-                                                                : 'text-gray-400',
-                                                        ]"
-                                                    />
-                                                    <template #panel="{ close }">
-                                                        <DatePicker
-                                                            v-model="monetizationDetails.termDate"
-                                                            is-required
-                                                            @close="close"
-                                                        />
-                                                    </template>
-                                                </UPopover>
-                                                <template #error="{ error }">
-                                                    <span
-                                                        :class="[
-                                                            error
-                                                                ? 'text-red-500 dark:text-red-400'
-                                                                : 'text-primary-500 dark:text-primary-400',
-                                                        ]"
-                                                    >
-                                                        {{ isPerpetual || monetizationDetails.termDate ? '' : error }}
-                                                    </span>
-                                                </template>
-                                            </UFormGroup>
-                                            <UFormGroup :label="$t('data.designer.perpetual')">
-                                                <UCheckbox
-                                                    :model-value="isPerpetual"
-                                                    class="mt-2.5 justify-center"
-                                                    @update:model-value="(value: boolean) => updatePerpetual(value)"
-                                                />
-                                            </UFormGroup>
-                                        </div>
-                                    </div>
-                                    <UFormGroup :label="$t('termsConditions')" name="extraTerms">
-                                        <UTextarea
-                                            v-model="monetizationDetails.extraTerms"
-                                            :rows="4"
-                                            :placeholder="$t('data.designer.typeTerms')"
-                                            resize
-                                        />
-                                    </UFormGroup>
-                                    <UFormGroup
-                                        :label="$t('data.designer.additionalRenewalTerms')"
-                                        name="additionalRenewalTerms"
-                                    >
-                                        <UTextarea
-                                            v-model="monetizationDetails.additionalRenewalTerms"
-                                            :rows="4"
-                                            :placeholder="$t('data.designer.additionalRenewalTerms')"
-                                            resize
-                                        />
-                                    </UFormGroup>
-                                    <div class="flex flex-row gap-4">
-                                        <div class="flex-1 flex gap-4">
-                                            <UFormGroup
-                                                :label="$t('data.designer.noticeForNonRenewal')"
-                                                class="flex-1"
-                                                required
-                                                name="nonRenewalDays"
-                                            >
-                                                <UInput
-                                                    v-model.number="monetizationDetails.nonRenewalDays"
-                                                    :placeholder="$t('data.designer.noticeForNonRenewal')"
-                                                    type="numeric"
-                                                >
-                                                    <template #trailing>
-                                                        <span class="text-gray-500 text-xs">{{ t('days') }}</span>
-                                                    </template>
-                                                </UInput>
-                                            </UFormGroup>
-                                        </div>
-                                        <div class="flex-1 flex gap-4">
-                                            <UFormGroup
-                                                :label="$t('data.designer.maximumDaysContractBreach')"
-                                                class="flex-1"
-                                                required
-                                                name="contractBreachDays"
-                                            >
-                                                <UInput
-                                                    v-model.number="monetizationDetails.contractBreachDays"
-                                                    :placeholder="$t('data.designer.maximumDaysContractBreach')"
-                                                    type="numeric"
-                                                >
-                                                    <template #trailing>
-                                                        <span class="text-gray-500 text-xs">{{ t('days') }}</span>
-                                                    </template>
-                                                </UInput>
-                                            </UFormGroup>
-                                        </div>
-                                    </div>
-                                    <div class="relative">
-                                        <div
-                                            class="absolute right-0 top-0 text-sm flex gap-4 items-center font-semibold"
-                                        >
-                                            <span class="text-gray-700">
-                                                {{ t('data.designer.hasPersonalData') }}
-                                            </span>
-                                            <UCheckbox
-                                                :model-value="hasPersonalData"
-                                                class="justify-center"
-                                                @update:model-value="(value: boolean) => updatePersonal(value)"
-                                            />
-                                        </div>
-                                        <UFormGroup
-                                            :label="$t('data.designer.personalDataTerms')"
-                                            name="personalDataTerms"
-                                            :required="hasPersonalData"
-                                        >
-                                            <UTextarea
-                                                v-model="monetizationDetails.personalDataTerms"
-                                                :rows="4"
-                                                :placeholder="$t('data.designer.personalDataTerms')"
-                                                resize
-                                                :disabled="!hasPersonalData"
-                                            />
-                                        </UFormGroup>
-                                    </div>
                                 </div>
                             </div>
                         </template>
@@ -970,23 +388,6 @@ async function onSubmit(): Promise<void> {
 
                         <template v-if="monetizationDetails.type === 'nft'">
                         </template> -->
-
-                        <!-- Terms Document -->
-                        <div v-show="monetizationDetails.license">
-                            <p class="font-semibold text-sm mt-5 mb-1.5">{{ $t('licenseDetails') }}</p>
-                            <DataShareTerms
-                                :asset-offering-details="assetOfferingDetails"
-                                :monetization-details="monetizationDetails"
-                                @update:contract-terms="(value: string) => updateContractTerms(value)"
-                            />
-                        </div>
-
-                        <div class="w-full flex items-center justify-between mt-4">
-                            <UButton size="md" color="gray" variant="outline" @click="emit('changePage', 0)">
-                                {{ $t('back') }}
-                            </UButton>
-                            <UButton size="md" type="submit" @click="onSubmit">{{ $t('next') }} </UButton>
-                        </div>
                     </UForm>
                 </Transition>
             </div>
