+++ i18n.config.ts
@@ -112,6 +112,17 @@ export default defineI18nConfig(() => ({
                         accessPoliciesEditor: 'Access Policies Editor',
                         preview: 'Preview',
                     },
+                    fields: {
+                        licenseSelection: 'License Selection',
+                        downloadLimitFrequency: 'Download Limit and Frequency',
+                        exclusivityAndAvailability: 'Exclusivity and Availability',
+                        licensingDetails: 'Licensing Details',
+                        additionalTerms: 'Additional Terms',
+                        contractDetails: 'Contract Details',
+                        personalData: 'Personal Data',
+                    },
+                    licenseDetails: 'License Details',
+                    licenseDetailsDescription: 'Details about the license you wish to apply to your asset',
                     noDatasetSelected: 'No Dataset Selected',
                     pleaseSelectFromCatalog: 'Please select a dataset from the catalogue to publish data.',
                     noTitle: 'No Title Given',
@@ -193,7 +204,7 @@ export default defineI18nConfig(() => ({
                     annual: 'Annual',
                     pleaseCheck: 'Please check the form and fill in all required (*) fields',
                     availability: 'Availability',
-                    regionCountry: 'Region / Country',
+                    regionCountry: 'Country',
                     worldwide: 'Worldwide',
                     transferable: 'Transferable',
                     nonTransferable: 'Non-transferable',
